requirements.txt
Django==5.2.9
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3